requirements.txt
astroid==2.4.1
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.12
django-filter==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
idna==2.9
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.5.2
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.12.1
